# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAText API
  version: '1.0'
  description: |-
    API per l'applicazione WASAText che consente di chattare con le persone o gruppi di persone.
    Gli endpoint che agiscono sull'utente autenticato (es. setMyUserName, setMyPhoto)
    non richiedono un User ID nel percorso, poiché l'identificativo viene estratto
    dall'header di autorizzazione della richiesta.
servers:
  - url: http://localhost:3000
security:
  - bearerAuth: []
...
paths:
  /session:
    post:
      tags: ["login"]
      summary: Logs in the user
      description: If the user does not exist, it will be created, and an identifier is returned. If the user exists, the user identifier is returned.
      operationId: doLogin
      requestBody:
        description: User details for login or registration
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Maria
                  minLength: 3
                  maxLength: 16
                  pattern: "^[a-zA-Z0-9_-]{3,16}$"
                  description: The user's desired username.
              required:
                - name
      responses:
        '201':
          description: User log-in action successful, returns the user ID
          content:
            application/json:
              schema:
                type: object
                properties:
                  identifier:
                    type: integer
                    example: 30
                    description: The unique identifier for the user.
...
  /groups:
    post:
      tags: ["groups"]
      summary: Creates a new group
      description: Creates a new group with an initial list of members.
      operationId: createGroup
      requestBody:
        description: Group name and initial members
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  $ref: "#/components/schemas/Name"
                memberIds:
                  type: array
                  items:
                    $ref: "#/components/schemas/Id"
                  minItems: 1
                  maxItems: 50
                  description: List of user IDs to add to the group.
              required:
                - name
                - memberIds
      responses:
        "201":
          description: Group created successfully, returns the group ID
          content:
            application/json:
              schema:
                type: object
                properties:
                  groupId:
                    $ref: "#/components/schemas/Id"
  /groups/{groupId}/name:
    parameters:
      - $ref: "#/components/parameters/groupId"
    put:
      tags: ["groups"]
      summary: Updates the name of a group
      description: Changes the name of the specified group.
      operationId: setGroupName
      requestBody:
        description: New name for the group
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  $ref: "#/components/schemas/Name"
              required:
                - name
      responses:
        "200":
          description: Name changed successfully
          content: {}
        "403":
          description: User is not a member of the group
        "404":
          description: Group not found
  /groups/{groupId}/photo:
    parameters:
      - $ref: "#/components/parameters/groupId"
    put:
      tags: ["groups"]
      summary: Updates the group profile photo
      description: Uploads a new profile photo for the group.
      operationId: setGroupPhoto
      requestBody:
        description: New group photo file
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
                  description: New group photo file (e.g., JPEG, PNG).
              required:
                - image
      responses:
        "200":
          description: Photo changed successfully
          content: {}
        "400":
          description: Invalid input, missing file or incorrect format
        "403":
          description: User is not a member of the group
        "404":
          description: Group not found
  /groups/{groupId}/members:
    parameters:
      - $ref: "#/components/parameters/groupId"
    post:
      tags: ["groups"]
      summary: Adds a list of users to a group
      description: Adds specified users to the existing group.
      operationId: addToGroup
      requestBody:
        description: List of user IDs to add
        content:
          application/json:
            schema:
              type: object
              properties:
                userIds:
                  type: array
                  items:
                    $ref: "#/components/schemas/Id"
                  minItems: 1
                  maxItems: 50
                  description: List of user IDs to add to the group.
              required:
                - userIds
      responses:
        "200":
          description: Users added successfully
          content: {}
        "403":
          description: User is not a member of the group
        "404":
          description: Group or one of the users not found
  /groups/{groupId}/members/{userId}:
    parameters:
      - $ref: "#/components/parameters/groupId"
      - $ref: "#/components/parameters/userId"
    delete:
      tags: ["groups"]
      summary: Makes a user leave a group
      description: Removes the specified user from the group.
      operationId: leaveGroup
      responses:
        "200":
          description: User left the group successfully
          content: {}
        "404":
          description: Group or user not found
        "403":
          description: User is not a member of the group
components:
  schemas:
    Id:
      description: A unique identifier (integer).
      type: integer
      example: 1234
    Name:
      description: A new username or group name, between 3 and 16 characters.
      type: string
      example: "Emanuele Panizzi"
      minLength: 3
      maxLength: 16
...
  parameters:
...
    groupId:
      name: groupId
      in: path
      required: true
      description: The unique identifier for a group.
      schema:
        $ref: "#/components/schemas/Id"
    userId:
      name: userId
      in: path
      required: true
      description: The unique identifier for a user.
      schema:
        $ref: "#/components/schemas/Id"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT